Cleaner module separation
